requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
feedparser
newspaper3k
lxml_html_clean
passlib[bcrypt]
python-jose[cryptography]
python-multipart
firebase-admin
